Increased clarity, fixed some logic, refactored code
